+++ events/index.js
@@ -1,23 +1,23 @@
 // This is where we `require` all of our event handlers ('event:name': handler)
 const debug = require('./debug');
 const connection = require('./connection');
-const gameLeave = require('./gameLeave.js');
 const gameCreation = require('./game-creation');
+const playerLeaveGame = require('./player-leave-game');
 const playerJoinGame = require('./player-join-game');
-const stationLogin = require('./station-login.js');
-const playerLogin = require('./playerLogin');
-const teamJoin = require('./teamJoin');
-const teamLeave = require('./teamLeave.js');
+const playerJoinTeam = require('./player-join-team');
+const playerLeaveTeam = require('./player-leave-team');
+const playerLogin = require('./player-login');
+const stationLogin = require('./station-login');
 
 // And then combine them into one object
 module.exports = {
   debug,
   connection,
   gameCreation,
-  gameLeave,
+  playerLeaveGame,
   playerJoinGame,
-  stationLogin,
+  playerJoinTeam,
+  playerLeaveTeam,
   playerLogin,
-  teamJoin,
-  teamLeave,
+  stationLogin,
 };
